PCHIP_module
Added separate module for PCHIP-based unfolds. Removed PCHIP functions from cspline. Added docstrings for PCHIP functions
